Disallow dots in beginning and end of usernames
The usernames are "regexd" as before, including dots in the beginning, end
and repeated within the usernames and then parsed in the parser to work
with the username conditions regarding dots in usernames.

Usernames can't start with a dot (invalid), can't end with a dot (parse
`@foo.` as `foo`) and can't include repeated dots (parse `@foo..bar` as
`foo`).